// src/types/searchBoxConfig.ts
export interface SearchBoxConfig {
  openInNewTab: boolean;
  clearAfterSearch: boolean;
  bgColor: string;
  focusBgColor: string;
  height: string;
  borderRadius: string;
  borderColor: string;
  paddingLeft: string;
  paddingRight: string;
  fontSize: string;
  textColor: string;
  iconSize: string;
}